requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
groq==0.9.0
openpyxl==3.1.2
pandas==2.1.4
requests==2.31.0
python-dateutil==2.8.2
